requirements.txt
requests>=2.31.0
scrapfly-sdk>=1.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0